Fix regression and two review-found bugs in disk scan/import path
- DiskScanService: ArtistScannedEvent now fires for all artists even
  when a scan batch's aggregate file list is empty. The early return
  added earlier for batching skipped this event entirely for every
  artist in a batch whenever even one folder had zero files.
- CandidateService: CommonProportion used integer division, silently
  truncating any partial fingerprint match to 0 and always failing
  the > 0.6 acceptance check unless every track matched exactly.
- UpgradeMediaFileService: the new file is now transferred before the
  old file is deleted/recycled, instead of always deleting first. The
  old file is only cleared preemptively on a destination collision
  (the common in-place-upgrade case, which requires it); for every
  other transfer failure (disk full, permissions, network share
  drop), the old file now survives instead of being lost with no
  replacement written.

+++ src/NzbDrone.Core/MediaFiles/TrackImport/Identification/CandidateService.cs
@@ -208,7 +208,7 @@ namespace NzbDrone.Core.MediaFiles.TrackImport.Identification
             {
                 Release = x,
                 TrackCount = x.TrackCount,
-                CommonProportion = x.Tracks.Value.Select(y => y.ForeignRecordingId).Intersect(recordingIds).Count() / localAlbumRelease.TrackCount
+                CommonProportion = (double)x.Tracks.Value.Select(y => y.ForeignRecordingId).Intersect(recordingIds).Count() / localAlbumRelease.TrackCount
             })
                 .Where(x => x.CommonProportion > 0.6)
                 .ToList()
